MAINT: with codes growing, split accessor.py to a package
