+++ doc/source/reference/accessor.rst
@@ -2,8 +2,28 @@ Accessor
 ========
 .. currentmodule:: dtoolkit.accessor
 
+
+Series Accessor
+-------------------
+.. autosummary::
+    :toctree: api/accessor/
+
+    DropInfSeriesAccessor
+
+
+DataFrame Accessor
+-------------------
+.. autosummary::
+    :toctree: api/accessor/
+
+    FilterInAccessor
+    RepeatAccessor
+
+
+Both Series and DataFrame Accessor
+-------------------
 .. autosummary::
     :toctree: api/accessor/
 
     ColumnAccessor
-    RepeatAccessor
+
