+++ dtoolkit/accessor/dataframe.py
@@ -1,88 +1,9 @@
 from __future__ import annotations
 
 import pandas as pd
-from numpy import inf
 from pandas.api.extensions import register_dataframe_accessor
-from pandas.api.extensions import register_series_accessor
 
-from ._typing import Pd
-
-__all__ = [
-    "ColumnAccessor",
-    "DropInfAccessor",
-    "FilterInAccessor",
-    "RepeatAccessor",
-]
-
-
-class Accessor:
-    """
-    :obj:`~pandas.Series` and :obj:`~pandas.DataFrame` objects base class.
-
-    Parameters
-    ----------
-    pd_obj : {:obj:`~pandas.Series`, :obj:`~pandas.DataFrame`}
-        pandas's object
-
-    See Also
-    --------
-    pandas.api.extensions.register_series_accessor
-    pandas.api.extensions.register_dataframe_accessor
-
-    Examples
-    --------
-    >>> from dtoolkit.accessor import Accessor
-    >>> from pandas.api.extensions import register_dataframe_accessor
-    >>> from pandas.api.extensions import register_series_accessor
-
-    Hook function to :obj:`~pandas.DataFrame` without any other function just
-    one (ex: :func:`~pandas.DataFrame.HookingName`)
-
-    >>> @register_series_accessor('HookingName')
-    ... class MyAccessor(Accessor):
-    ...     def __call__(self):
-    ...         ...
-    """
-
-    def __init__(self, pd_obj: Pd):
-        self.pd_obj = pd_obj
-
-
-@register_dataframe_accessor("cols")
-@register_series_accessor("cols")
-class ColumnAccessor(Accessor):
-    """
-    A API to gather :attr:`~pandas.Series.name` and
-    :attr:`~pandas.DataFrame.columns` to one.
-
-    See Also
-    --------
-    pandas.Series.name
-    pandas.DataFrame.columns
-    """
-
-    def __call__(self) -> str | list[str]:
-        if isinstance(self.pd_obj, pd.Series):
-            return self.pd_obj.name
-
-        return self.pd_obj.columns.tolist()
-
-
-@register_dataframe_accessor("dropinf")
-@register_series_accessor("dropinf")
-class DropInfAccessor(Accessor):
-    def __call__(self, inplace: bool = False) -> Pd | None:
-        mask = ~self.pd_obj.isin([inf, -inf])
-        if isinstance(self.pd_obj, pd.DataFrame):
-            mask = mask.all(axis=1)
-
-        result = self.pd_obj[mask]
-
-        if inplace:
-            self.pd_obj._update_inplace(result)
-            return None
-
-        return result
+from .base import Accessor
 
 
 @register_dataframe_accessor("filterin")
